[fixup] Namespacing of DTO and events filter for membership creates
